API_Gateway/requirements.txt
# Flask
flask
python-dotenv
python-environ

# Gunicorn
gunicorn

# Requests
requests